Simple class prototype to avoid messages about unknown classes from setMethod

For SQL Server, conn@quote will return the quotation mark, however both quotation marks as well as square bracket are used interchangeably for delimited identifiers. See: https://learn.microsoft.com/en-us/sql/relational-databases/databases/database-identifiers?view=sql-server-ver16 Therefore strip the brackets first, and then call the DBI method that strips the quotation marks. TODO: the generic implementation in DBI should take a quote char as parameter.

Local temp tables are stored as [tempdb].[dbo].[#name]________(padding using underscores)[numeric identifier]

If we can identify that the name is that of a local temp table then adjust the identifier and query appropriately.

Will warn user if temporary is set to TRUE but table name does not conform to local temp table naming conventions. If writing to a global temp table, user should not set the temporary flag to TRUE.

Arguments

conn

A DBIConnection object, as returned by dbConnect().

x

An SQL or Id object.

...

Other arguments passed on to methods.

name

The table name, passed on to dbQuoteIdentifier(). Options are:

a character string with the unquoted DBMS table name, e.g. "table_name",

a call to Id() with components to the fully qualified table name, e.g. Id(schema = "my_schema", table = "table_name")

a call to SQL() with the quoted and fully qualified table name given verbatim, e.g. SQL('"my_schema"."table_name"')

con

A database connection.

table

The table name, passed on to dbQuoteIdentifier(). Options are:

a character string with the unquoted DBMS table name, e.g. "table_name",

a call to Id() with components to the fully qualified table name, e.g. Id(schema = "my_schema", table = "table_name")

a call to SQL() with the quoted and fully qualified table name given verbatim, e.g. SQL('"my_schema"."table_name"')

fields

Either a character vector or a data frame.

A named character vector: Names are column names, values are types. Names are escaped with dbQuoteIdentifier(). Field types are unescaped.

A data frame: field types are generated using dbDataType().

row.names

Either TRUE, FALSE, NA or a string.

If TRUE, always translate row names to a column called "row_names". If FALSE, never translate row names. If NA, translate rownames only if they're a character vector.

A string is equivalent to TRUE, but allows you to override the default name.

For backward compatibility, NULL is equivalent to FALSE.

temporary

If TRUE, will generate a temporary table statement.

Value

dbExistsTable() returns a logical scalar, TRUE if the table or view specified by the name argument exists, FALSE otherwise.

This includes temporary tables if supported by the database.

Details

True if:

If catalog_name is supplied it must equal "temdb" or "%" ( wildcard )

Name must start with "#" followd by a non-"#" character

Note, the implementation here is such that it assumes the metadata attribute is set such that catalog functions accept wildcard entries.

Driver note. OEM driver will return correctly for name, catalog_name = "tempdb" in some circumstances. For exmaple if the name has no underscores to beginwith. FreeTDS, will not index the table correctly unless name is adjusted ( allowed trailing wildcards to accomodate trailing underscores and postfix ).

Therefore, in all cases query for name___%.

In both cases a simple CREATE TABLE statement is used / the table identifier is the differentiator ( viz-a-viz creating a non-temp table ).

Failure modes

An error is raised when calling this method for a closed or invalid connection. An error is also raised if name cannot be processed with dbQuoteIdentifier() or if this results in a non-scalar.

Specification

The name argument is processed as follows, to support databases that allow non-syntactic names for their objects:

If an unquoted table name as string: dbExistsTable() will do the quoting, perhaps by calling dbQuoteIdentifier(conn, x = name)

If the result of a call to dbQuoteIdentifier(): no more quoting is done

For all tables listed by dbListTables(), dbExistsTable() returns TRUE.

See also

Examples

con <- dbConnect(RSQLite::SQLite(), ":memory:")

dbExistsTable(con, "iris")
#> [1] FALSE
dbWriteTable(con, "iris", iris)
dbExistsTable(con, "iris")
#> [1] TRUE

dbDisconnect(con)