dbConnect,OdbcDriver-method {odbc}	R Documentation

Connect to a ODBC compatible database

Description

Connect to a ODBC compatible database

Usage

## S4 method for signature 'OdbcDriver'
dbConnect(drv, dsn = NULL, ..., timezone = "UTC",
  encoding = "", bigint = c("integer64", "integer", "numeric", "character"),
  timeout = 10, driver = NULL, server = NULL, database = NULL,
  uid = NULL, pwd = NULL, dbms.name = NULL, .connection_string = NULL)

Arguments

drv
an object that inherits from DBIDriver, or an existing DBIConnection object (in order to clone an existing connection).

dsn
The Data Source Name.

...
Additional ODBC keywords, these will be joined with the other arguments to form the final connection string.

timezone
The Server time zone. Useful if the database has an internal timezone that is not 'UTC'. If the database is in your local timezone set to Sys.timezone(). See OlsonNames() for a complete list of available timezones on your system.

encoding
The text encoding used on the Database. If the database is not using UTF-8 you will need to set the encoding to get accurate re-encoding. See iconvlist() for a complete list of available encodings on your system. Note strings are always returned UTF-8 encoded.

bigint
The R type that SQL_BIGINT types should be mapped to, default is bit64::integer64, which allows the full range of 64 bit integers.

timeout
Time in seconds to timeout the connection attempt. Setting a timeout of Inf indicates no timeout. (defaults to 10 seconds).

driver
The ODBC driver name.

server
The server hostname.

database
The database on the server.

uid
The user identifier.

pwd
The password to use.

dbms.name
The database management system name. This should normally be queried automatically by the ODBC driver. This name is used as the class name for the OdbcConnect object returned from dbConnect(). However if the driver does not return a valid value it can be set manually with this parameter.

.connection_string
A complete connection string, useful if you are copy pasting it from another source. If this argument is used any additional arguments will be appended to this string.

Details

The connection string keywords are driver dependent. The parameters documented here are common, but some drivers may not accept them. Please see the specific driver documentation for allowed parameters, https://www.connectionstrings.com is also a useful resource of example connection strings for a variety of databases.

[Package odbc version 1.1.5 Index]