dbListFields,OdbcConnection,character-method {odbc}	R Documentation

List field names of a remote table

Description

List field names of a remote table

Usage

## S4 method for signature 'OdbcConnection,character'
dbListFields(conn, name,
  catalog_name = NULL, schema_name = NULL, column_name = NULL, ...)

Arguments

conn
A DBIConnection object, as returned by dbConnect().

name
a character string with the name of the remote table.

catalog_name
The name of the catalog to return, the default returns all catalogs.

schema_name
The name of the schema to return, the default returns all schemas.

column_name
The name of the column to return, the default returns all columns.

...
Other parameters passed on to methods.

Details

% can be used as a wildcard in any of the search parameters to match 0 or more characters. _ can be used to match any single character.

Value

a character vector

See Also

dbColumnInfo() to get the type of the fields.

Other DBIConnection generics: DBIConnection-class, dbDataType, dbDisconnect, dbExecute, dbExistsTable, dbGetException, dbGetInfo, dbGetQuery, dbIsValid, dbListObjects, dbListResults, dbListTables, dbReadTable, dbRemoveTable, dbSendQuery, dbSendStatement, dbWriteTable

[Package odbc version 1.1.5 Index]