dbListTables,OdbcConnection-method {odbc}	R Documentation

List remote tables

Description

Returns the unquoted names of remote tables accessible through this connection. This should include views and temporary objects, but not all database backends (in particular RMariaDB and RMySQL) support this.

Methods in other packages

odbc::dbListTables("OdbcConnection")

odbc::dbListTables("Teradata")

Usage

## S4 method for signature 'OdbcConnection'
dbListTables(conn, catalog_name = NULL,
  schema_name = NULL, table_name = NULL, table_type = NULL, ...)

Arguments

conn
A DBIConnection object, as returned by dbConnect().

catalog_name
The name of the catalog to return, the default returns all catalogs.

schema_name
The name of the schema to return, the default returns all schemas.

table_name
The name of the table to return, the default returns all tables.

table_type
The type of the table to return, the default returns all table types.

...
Other parameters passed on to methods.

Details

% can be used as a wildcard in any of the search parameters to match 0 or more characters. _ can be used to match any single character.

Value

dbListTables() returns a character vector that enumerates all tables and views in the database. Tables added with dbWriteTable() are part of the list, including temporary tables if supported by the database. As soon a table is removed from the database, it is also removed from the list of database tables.

The returned names are suitable for quoting with dbQuoteIdentifier(). An error is raised when calling this method for a closed or invalid connection.

Additional arguments

TBD: temporary = NA

This must be provided as named argument. See the "Specification" section for details on their usage.

See Also

The ODBC documentation on Pattern Value Arguments for further details on the supported syntax.

[Package odbc version 1.1.5 Index]