odbcConnectionActions {odbc}	R Documentation

List the actions supported for the connection

Description

Return a list of actions that can be performed on the connection.

Usage

odbcConnectionActions(connection)

Arguments

connection
A connection object, as returned by dbConnect().

Details

The list returned is a named list of actions, where each action has the following properties:

callback

A function to be invoked to perform the action

icon

An optional path to an icon representing the action

Value

A named list of actions that can be performed on the connection.

[Package odbc version 1.1.5 Index]