odbcPreviewObject {odbc}	R Documentation

Preview the data in an object.

Description

Return the data inside an object as a data frame.

Usage

odbcPreviewObject(connection, rowLimit, ...)

Arguments

connection
A connection object, as returned by dbConnect().

rowLimit
The maximum number of rows to display.

...
Parameters specifying the object.

Details

The object to previewed must be specified as one of the arguments (e.g. table = "employees"); depending on the driver and underlying data store, additional specification arguments may be required.

Value

A data frame containing the data in the object.

[Package odbc version 1.1.5 Index]